backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
openai>=1.30.0
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1
websockets==12.0
